// src/app/model/contact.model.ts
export class Contact {
    id:number;
    firstName: any;
    lastName: any;
    email: any;
    mobile: any;
    image: any;
}

// export class Contact {
//     constructor(
//       public id: number,
//       public firstName: string,
//       public lastName: string,
//       public email: string,
//       public mobile: string,
//       public image: string
//     ) {}
//   }